+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 - add numeric character escapes: `\+0030` (#181)
 - add nested diagnostic locations in node CLI (#51)
 - fix duplicate output for `--no-color` option
+- fix performance and memory leak issues related to `regex`
 - also add internal `__typeof` directive and test types of directives
 
 **Full Changelog**: